Shutdown threads cleanly
Stop the kafka consumers first, which causes the stream iterators to
return false for has next and commits offsets for us. Returning false
from the iterator causes the doseq to exit and so awaiting termination
of the thread pool does not always wait the full time before
terminating.

We shut down the consumers in parallel now which seems to mitigate some
rebalances and slow shutdowns but every rebalance takes ~8 seconds
locally which is unfortunate because it makes test shutdown take a
while. If we only had one consumer across many threads this would behave
a lot better but then when committing offsets, you could commit an
offset that has been read but not processed. The slow shutdown seems
like the lesser evil.