Catch failing to commit offsets when shutting down
In the kafka consumer shutdown, it nils out its ZK client so the
kafka consumer still exists, but is unable to do certain ZK things. This
leads to a null pointer for the zk client while the kafka consumer has
shut down but the thread is still trying to commit offsets.

We can just ignore these strange exceptions from committing offsets when
shutting down and we may see a few of these messages again. We only want
to squash the errors when shutting down though otherwise normal
operation is compromised. Ideally, we wouldn't have to yank the consumer
away from the task consuming messages but that's the cleanest way to
break the blocking iterator on messages. Alternatively, we can do our
own poll with timeout as we look for next messages and also check if we
should be shutting down which would avoid the client being deactivated
suddenly but that's quite a bit more work.