Fix get-messages mock helper to keep poll in small intervals
The call to poll is supposed to be frequent so that kafka consumers can
synchronize on things like rebalances and partition assignments (that's how the
real kafka works). Allowing poll to become long with users that pass in long
timeouts is getting ready for issues since these consumers are going to be
rebalanced, etc, if there isn't high enough message volume to keep the consumers
poll looping anyways.